scripts/requirements.txt
Pillow==10.4.0
pillow-heif==0.18.0
